+++ src/main/java/playlist_abr/PlaylistModifyOutputBoundary.java
@@ -1,4 +1,4 @@
-package PlaylistABR;
+package playlist_abr;
 
 public interface PlaylistModifyOutputBoundary {
 }
