+++ src/main/java/playlist_abr/PlaylistModifyRequestModel.java
@@ -1,4 +1,4 @@
-package PlaylistABR;
+package playlist_abr;
 
 public class PlaylistModifyRequestModel {
 }
