+++ src/main/java/playlist_abr/PlaylistModifyResponseModel.java
@@ -1,4 +1,4 @@
-package PlaylistABR;
+package playlist_abr;
 
 public class PlaylistModifyResponseModel {
 }
