+++ src/main/java/playlist_screen/PlaylistModifyControl.java
@@ -1,4 +1,4 @@
-package PlaylistScreen;
+package playlist_screen;
 
 public class PlaylistModifyControl {
 }
