+++ src/main/java/playlist_screen/PlaylistModifyViewModel.java
@@ -1,4 +1,4 @@
-package PlaylistScreen;
+package playlist_screen;
 
 public class PlaylistModifyViewModel {
 }
