+++ src/main/java/playlist_screen/RecordPresenter.java
@@ -1,8 +1,8 @@
-package PlaylistScreen;
+package playlist_screen;
 
-import PlaylistABR.PlaylistCreateOutputBoundary;
-import PlaylistABR.PlaylistCreateResponseModel;
-import PlaylistEntities.Playlist;
+import playlist_abr.PlaylistCreateOutputBoundary;
+import playlist_abr.PlaylistCreateResponseModel;
+import playlist_entities.Playlist;
 
 public class RecordPresenter implements PlaylistCreateOutputBoundary {
     public PlaylistCreateResponseModel createRMConverter(Playlist playlist){
